untested: COMMIT now groks escaped filenames
This is nearly untested.

Also "Added and Modified" is detected and printed.
Note that it is complex to checkin staged data file-by-file
in the case that it is modified, as giving it on commandline
then also adds the newer changes, too.
Hence I leave this for future.  (Perhaps I oversaw an option
how to do so.)