git bring [file [revs]]: inverse to git fetch file
"git bundle create file revs" is a bit too unintuitive.

"git bring /file" provides the file which can be read by
"git fetch /file" again.  Easy and very symmetric.

By default this exports everything, that is not on upstream yet.
So "git bring" creates a file in /tmp/ which includes all
you probably want to push (but perhaps cannot because you
are on the wrong machine).

Easy and convenient.