git adate [seconds]: amend the commit date
This changes the commit date of the last commit to the given
seconds since the epoch.  This uses bash's printf.
Default is -1 which means current time.
Use -2 for the time the shell session begun.

Why?

This way you can define, where a commit in git-log is placed.
By using 0, automatically created commits will be placed
on the bottom (so they do not clutter up your recent history).
This is the main use, as well as to "touch" (bring up)
some commit to the top (which is the default).